Part 3
I planned out and implemented the final part of the assignment (part 3). This is the remove row function which needed to do multiple things, including changing the total size of the database based on how many rows were used and moving all rows after the deleted one up in the database.