Fixing show row function
As per the assignment brief the id and year fields of each record (database entry), when printed need to take up 6 and 4 spaces respectively. In this version, I tried to address that by truncating the number if it was too long and adding extra blank space if it was too short.